% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{as.matrix.subdist}
\alias{as.matrix.subdist}
\title{Coerce a \code{subdist} object to a matrix}
\usage{
\method{as.matrix}{subdist}(x, ...)
}
\arguments{
\item{x}{An object of class \code{subdist} as generated by [\verb{[.dist}].}

\item{...}{Additional arguments passed to \code{\link{as.matrix}}. Currently ignored.}
}
\value{
A numeric matrix storing the pairwise distances between the requested
observations.
}
\description{
Coerce a \code{subdist} object to a matrix.
}
\examples{
D <- stats::dist(iris[, 1:4])
x <- D[1:3, 2:7]
as.matrix(x)
}
