% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist}
\alias{dist}
\title{Distance Matrix Computation}
\usage{
dist(x, FUN = function(.x, .y) sqrt(sum(.x - .y)^2), labels = NULL, ...)
}
\arguments{
\item{x}{A list of objects of the same class.}

\item{FUN}{A function used to compute the distances between observations in
sample \code{x}.}

\item{labels}{A character vector of labels for the rows of the distance
matrix. If \code{NULL}, the row names of \code{x} are used if they exist. Otherwise,
integer labels are used.}

\item{...}{Additional arguments passed to \code{FUN}.}
}
\value{
A \code{\link[stats:dist]{stats::dist}} object.
}
\description{
Computes and returns the distance matrix computed by using the
specified distance measure to compute the distances between the rows of a
data matrix.
}
\examples{
cars |>
  apply(1, \(.x) .x, simplify = FALSE) |> # Convert to a list of observations
  dist()
}
