% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distops-package.R
\docType{package}
\name{distops-package}
\alias{distops}
\alias{distops-package}
\title{distops: Usual Operations for Distance Matrices in R}
\description{
It provides the subset operator for dist objects and a function to compute medoid(s) that are fully parallelized leveraging the 'RcppParallel' package. It also provides functions for package developers to easily implement their own parallelized dist() function using a custom 'C++'-based distance function.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lmjl-alea/distops}
  \item \url{https://lmjl-alea.github.io/distops/}
  \item Report bugs at \url{https://github.com/lmjl-alea/distops/issues}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID})

}
