% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_package.R
\name{use_distops}
\alias{use_distops}
\title{Setups package to use the \strong{distops} package}
\usage{
use_distops()
}
\value{
Nothing.
}
\description{
This function setups the package to use the \strong{distops} package.
It first creates the \code{DESCRIPTION} file adding the \strong{Rcpp} and
\strong{RcppParallel} packages to both the \verb{Imports:} and \verb{LinkingTo:} fields
and the \strong{distops} package to the \verb{LinkingTo:} field. It also adds the
\verb{SystemRequirements: GNU make} field. It then creates the \code{NAMESPACE} file
adding the \code{importFrom()} directives for the \strong{Rcpp} and \strong{RcppParallel}
packages and the \code{useDynLib()} directive for packages with compiled code.
It finally creates the \code{src/Makevars} and \code{src/Makevars.win} files with the
appropriate compilation flags.
}
\examples{
\donttest{
use_distops()
}
}
