% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distops-package.R
\docType{package}
\name{distops-package}
\alias{distops}
\alias{distops-package}
\title{distops: Usual Operations for Distance Matrices in R}
\description{
It implements an alternative `dist()` function that integrates parallelization and progress bars via the `future` framework. It also provides functions to manipulate distance matrices, such as an implementation of the subset operator which never turn the `dist` object into a full matrix.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lmjl-alea/distops}
  \item \url{https://lmjl-alea.github.io/distops/}
  \item Report bugs at \url{https://github.com/lmjl-alea/distops/issues}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID})

}
